Move Connections to Payment::Connection
