+++ config/routes.rb
@@ -7,9 +7,9 @@
 # https://guides.rubyonrails.org/routing.html
 Rails.application.routes.draw do
   # Legacy routes based off OAuth connections. We will update our OAuth providers information, but need these until we do.
-  get 'publishers/uphold_verified', to: 'connections/uphold_connections#edit'
-  get 'publishers/gemini_connection/new', to: 'connections/gemini_connections#edit'
-  get 'publishers/paypal_connections/connect_callback', to: 'connections/paypal_connections#connect_callback'
+  get 'publishers/uphold_verified', to: 'payment/connection/uphold_connections#edit'
+  get 'publishers/gemini_connection/new', to: 'payment/connection/gemini_connections#edit'
+  get 'publishers/paypal_connections/connect_callback', to: 'payment/connection/paypal_connections#connect_callback'
 
   # Routes for Browser Users to login via Uphold
   namespace :uphold_connections do
@@ -24,7 +24,7 @@ Rails.application.routes.draw do
   end
 
   # These routes are for connecting to 3rd-party payment providers.
-  namespace :connections do
+  namespace :connection, module: 'payment/connection' do
     resource :currency, only: [:show, :update]
     resource :stripe_connection
     resource :gemini_connection
